+++ tests/testthat/test-clonalOverlap.R
@@ -19,6 +19,14 @@ test_that("clonalOverlap works", {
   )
   
   expect_doppelganger( # warning from testthat: Removed 36 rows containing missing values (`geom_text()`).
+    "clonalOverlap_order_plot",
+    clonalOverlap(
+      combined, 
+      method = "raw",
+      order.by = c("P17B","P18B","P19B","P20B","P17L","P18L","P19L","P20L"))
+  )
+  
+  expect_doppelganger( # warning from testthat: Removed 36 rows containing missing values (`geom_text()`).
     "clonalOverlap_morisita_plot",
     clonalOverlap(
       combined,